+++ src/context/AuthContext.tsx
@@ -1,0 +1,10 @@
+import { createContext } from "react";
+
+type AuthContextType = {
+    user: any | null;
+    loading: boolean;
+    isAdmin: boolean;
+    isFriend: boolean;
+};
+
+export const AuthContext = createContext<AuthContextType | undefined>(undefined);
